requirements.txt
blinker==1.8.2
chardet==5.2.0
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.3
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.1
numpy==2.0.2
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
reportlab==4.2.5
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.4
zipp==3.20.2
